Added Exporter
Added the ability to export the time frame from the ScaleInfo page.
Moved all database reader-writer (just the mongo RW) to one file so that there is an abstract database class.
Minimized the effect on the database when the HX711 chip fails and returns a bad value (all 1s or all 0s).